+++ pg_search/src/index/writer.rs
@@ -44,7 +44,8 @@ use tantivy::{
 use thiserror::Error;
 
 use super::directory::{SearchDirectoryError, SearchFs, WriterDirectory};
-use crate::postgres::storage::{SegmentWriter, TantivyMetaDirectory};
+use crate::postgres::storage::atomic::AtomicDirectory;
+use crate::postgres::storage::segment_writer::SegmentWriter;
 
 /// Returns true if the file is "managed".
 /// Non-managed file are not subject to garbage collection.
@@ -89,15 +90,13 @@ impl Directory for BlockingDirectory {
     }
 
     fn open_write(&self, path: &Path) -> result::Result<WritePtr, OpenWriteError> {
-        pgrx::info!("open_write {:?}", path);
         Ok(io::BufWriter::new(Box::new(unsafe {
             SegmentWriter::new(self.relation_oid, path)
         })))
     }
 
     fn atomic_write(&self, path: &Path, data: &[u8]) -> io::Result<()> {
-        pgrx::info!("atomic_write {:?}", path);
-        let directory = unsafe { TantivyMetaDirectory::new(self.relation_oid) };
+        let directory = unsafe { AtomicDirectory::new(self.relation_oid) };
         match path.to_str().unwrap().ends_with("meta.json") {
             true => unsafe { directory.write_meta(data) },
             false => unsafe { directory.write_managed(data) },
@@ -108,7 +107,7 @@ impl Directory for BlockingDirectory {
 
     fn atomic_read(&self, path: &Path) -> result::Result<Vec<u8>, OpenReadError> {
         pgrx::info!("atomic_read: {:?}", path);
-        let directory = unsafe { TantivyMetaDirectory::new(self.relation_oid) };
+        let directory = unsafe { AtomicDirectory::new(self.relation_oid) };
         let data = match path.to_str().unwrap().ends_with("meta.json") {
             true => unsafe { directory.read_meta() },
             false => unsafe { directory.read_managed() },
