+++ pg_search/src/postgres/mod.rs
@@ -24,13 +24,13 @@ mod insert;
 pub mod options;
 mod range;
 mod scan;
-pub mod storage;
 mod vacuum;
 mod validate;
 
 pub mod customscan;
 pub mod datetime;
 pub mod index;
+pub mod storage;
 pub mod transaction;
 pub mod types;
 pub mod utils;
