+++ pg_search/src/postgres/storage/mod.rs
@@ -1,0 +1,4 @@
+pub mod atomic;
+pub mod buffer;
+pub mod metadata;
+pub mod segment_writer;
